requirements.txt
Django>=5.0,<6.0
gunicorn>=22.0
psycopg2-binary>=2.9
django-environ>=0.11
whitenoise>=6.7
Pillow>=10.0
djangorestframework>=3.15
